Add get_sources mehod to LexicalGraphINdex

+++ lexical-graph/src/graphrag_toolkit/lexical_graph/storage/vector/pg_vector_indexes.py
@@ -374,8 +374,8 @@ class PGIndex(VectorIndex):
                             value text,
                             metadata jsonb,
                             embedding vector({self.dimensions}),
-                            valid_from INTEGER DEFAULT -1,
-                            valid_to INTEGER DEFAULT -1
+                            valid_from BIGINT DEFAULT -1,
+                            valid_to BIGINT DEFAULT -1
                             );'''
                         )
                     except UniqueViolation:
